+++ finble-backend/src/main/java/com/ssafy/BlueMarble/websocket/dto/payload/game/UseDicePayload.java
@@ -17,6 +17,6 @@ public class UseDicePayload {
     private int currentPosition;                // 주사위 사용 후 이동하게 되는 위치
     private int salaryBonus;                    // 만약에 시작 위치를 통과했다면 월급 지급 아니라면 0
     private boolean canBuyLand;                 // 이동한 땅을 살 수 있는지 true, false
-    private int tollAmount;                     // 상대방의 땅이라면 내야하는 톨비
+    private Long tollAmount;                     // 상대방의 땅이라면 내야하는 톨비
     private ConstructPayload.Asset updatedAsset; // 주사위 던진 사람의 자산 정보 변동내역
 }
